Add test for git task
Split pipeline
