+++ pom.xml
@@ -37,6 +37,11 @@
 			<artifactId>jackson-databind</artifactId>
 			<version>2.15.2</version>
 		</dependency>
+		<dependency>
+			<groupId>org.eclipse.jgit</groupId>
+			<artifactId>org.eclipse.jgit</artifactId>
+			<version>6.6.0.202305301015-r</version>
+		</dependency>
 	</dependencies>
 
 	<build>
