--- src/test/java/ru/naumen/practiceTest/task1/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Тестирование первого задания
- * @author aspirkin
- */
-package ru.naumen.practiceTest.task1;
