+++ src/test/java/ru/naumen/practiceTest/task2/package-info.java
@@ -1,5 +1,5 @@
 /**
- * Тестирование второго задания
+ * Тестирование первого задания
  * @author aspirkin
  */
 package ru.naumen.practiceTest.task2;
