+++ src/test/java/ru/naumen/practiceTest/task3/package-info.java
@@ -1,5 +1,5 @@
 /**
- * Тестирование третьего задания
+ * Тестирование второго задания
  * @author aspirkin
  */
 package ru.naumen.practiceTest.task3;
