+++ src/test/java/ru/naumen/practiceTest/task4/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Тестирование третьего задания
+ * @author aspirkin
+ */
+package ru.naumen.practiceTest.task4;
